feat: Adding TaskPage on views path
